refactor: 총 상금 구하는 로직을 더 적절한 LottoDrawingResult에게 책임 부여
