+++ src/main/kotlin/domain/LottoMachine.kt
@@ -1,6 +1,5 @@
 package domain
 
-import domain.model.LottoDrawingResult
 import domain.model.Margin
 import domain.model.Money
 
@@ -18,10 +17,6 @@ class LottoMachine(
 
     fun calculateMargin(prize: Money): Margin {
         return Margin(prize.amount * LOTTO_PRICE / money.amount / LOTTO_PRICE.toDouble())
-    }
-
-    fun calculateTotalPrize(result: LottoDrawingResult): Money {
-        return Money(result.statistics.entries.sumOf { (rank, count) -> rank.winningMoney * count.toLong() })
     }
 
     companion object {
