+++ src/main/kotlin/domain/model/LottoDrawingResult.kt
@@ -2,4 +2,9 @@ package domain.model
 
 import domain.Rank
 
-data class LottoDrawingResult(val statistics: Map<Rank, Int>)
+data class LottoDrawingResult(val statistics: Map<Rank, Int>) {
+
+    fun calculateTotalPrize(): Money {
+        return Money(statistics.entries.sumOf { (rank, count) -> rank.winningMoney * count.toLong() })
+    }
+}
